Move MiriamValidator and use it in controller
