+++ src/biomappings/tests/test_validity.py
@@ -7,48 +7,11 @@ from collections import Counter
 import requests
 
 from biomappings import load_false_mappings, load_mappings, load_predictions
-from biomappings.utils import get_canonical_tuple
+from biomappings.utils import get_canonical_tuple, MiriamValidator
 
 mappings = load_mappings()
 predictions = load_predictions()
 incorrect = load_false_mappings()
-
-
-class InvalidPrefix(ValueError):
-    """Raised for an invalid prefix."""
-
-
-class InvalidIdentifier(ValueError):
-    """Raised for an invalid identifier."""
-
-
-class MiriamValidator:
-    """Validate prefix/identifier pairs based on the MIRIAM database."""
-
-    def __init__(self):  # noqa: D107
-        self.entries = self._load_identifiers_entries()
-
-    @staticmethod
-    def _load_identifiers_entries():
-        url = 'https://registry.api.identifiers.org/resolutionApi/getResolverDataset'
-        res = requests.get(url)
-        regj = res.json()
-        patterns = {
-            entry['prefix']: {
-                'pattern': re.compile(entry['pattern']),
-                'namespace_embedded': entry['namespaceEmbeddedInLui'],
-            }
-            for entry in sorted(regj['payload']['namespaces'], key=lambda x: x['prefix'])
-        }
-        return patterns
-
-    def check_valid_prefix_id(self, prefix, identifier):
-        """Check the prefix/identifier pair is valid."""
-        if prefix not in self.entries:
-            raise InvalidPrefix(prefix)
-        entry = self.entries[prefix]
-        if not re.match(entry['pattern'], identifier):
-            raise InvalidIdentifier(identifier)
 
 
 miriam_validator = MiriamValidator()
