+++ src/biomappings/utils.py
@@ -2,7 +2,9 @@
 
 """Utilities."""
 
+import re
 import os
+import requests
 from subprocess import CalledProcessError, check_output  # noqa: S404
 from typing import Any, Mapping, Optional, Tuple
 
@@ -56,3 +58,44 @@ def get_canonical_tuple(mapping: Mapping[str, Any]) -> Tuple[str, str, str, str]
     if source > target:
         source, target = target, source
     return (*source, *target)
+
+
+class InvalidPrefix(ValueError):
+    """Raised for an invalid prefix."""
+
+
+class InvalidIdentifier(ValueError):
+    """Raised for an invalid identifier."""
+
+
+class MiriamValidator:
+    """Validate prefix/identifier pairs based on the MIRIAM database."""
+
+    def __init__(self):  # noqa: D107
+        self.entries = self._load_identifiers_entries()
+
+    @staticmethod
+    def _load_identifiers_entries():
+        url = 'https://registry.api.identifiers.org/resolutionApi/getResolverDataset'
+        res = requests.get(url)
+        regj = res.json()
+        patterns = {
+            entry['prefix']: {
+                'pattern': re.compile(entry['pattern']),
+                'namespace_embedded': entry['namespaceEmbeddedInLui'],
+            }
+            for entry in sorted(regj['payload']['namespaces'], key=lambda x: x['prefix'])
+        }
+        return patterns
+
+    def namespace_embedded(self, prefix):
+        """Return True if the namespace is embedded for the given prefix."""
+        return self.entries[prefix]['namespace_embedded']
+
+    def check_valid_prefix_id(self, prefix, identifier):
+        """Check the prefix/identifier pair is valid."""
+        if prefix not in self.entries:
+            raise InvalidPrefix(prefix)
+        entry = self.entries[prefix]
+        if not re.match(entry['pattern'], identifier):
+            raise InvalidIdentifier(identifier)
